Ensure that |partial_common_block| selects the actual Bruhat interval below
Before, we just ensured the interval was generated and returned the whole common
block containing the original parameter. Now we generate and report the subset,
then select its top element to retuen from the partial block instance of
|Rep_table::lookup|, after which the wrapper function for |partial_common_block|
reconstructs the interval using the |hasse| vectors of stored Bruhat order.

+++ sources/gkmod/blocks.h
@@ -120,7 +120,7 @@ public:
   KGBElt x(BlockElt z) const { assert(z<size()); return info[z].x; }
   KGBElt y(BlockElt z) const { assert(z<size()); return info[z].y; }
 
-  size_t length(BlockElt z) const { return info[z].length; }
+  unsigned short length(BlockElt z) const { return info[z].length; }
 
   // first element of length (at least) |l|, or |size()| if there are none
   BlockElt length_first(size_t l) const;
